% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{get_geo_boundary_data}
\alias{get_geo_boundary_data}
\title{get_geo_boundary_data()}
\usage{
get_geo_boundary_data(type = "lsoa")
}
\arguments{
\item{type}{Options are \code{lsoa}, \code{la} (local authority district/unitary authority) and \code{lpa} (local planning authority); \code{lsoa} by default.}
}
\description{
get_geo_boundary_data()
}
