% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jts_tables}
\alias{jts_tables}
\title{Metadata on JTS tables}
\format{
A data frame with 223 rows and 7 variables:
\describe{
\item{\code{table_code}}{character COLUMN_DESCRIPTION}
\item{\code{csv_file_name}}{character COLUMN_DESCRIPTION}
\item{\code{table_title}}{character COLUMN_DESCRIPTION}
\item{\code{table_url}}{character COLUMN_DESCRIPTION}
\item{\code{csv_url}}{character COLUMN_DESCRIPTION}
\item{\code{service}}{character COLUMN_DESCRIPTION}
\item{\code{year}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
jts_tables
}
\description{
This dataset reports available Journey Time Statistics files
}
\details{
DETAILS
}
\keyword{datasets}
